backend/requirements.txt
fastapi
uvicorn
google-generativeai
qdrant-client
asyncpg
sqlalchemy
python-dotenv
pydantic-settings
alembic
httpx
